+++ ext_localconf.php
@@ -106,6 +106,8 @@ $boot = static function () {
     // FAL upgrade wizards
     $GLOBALS['TYPO3_CONF_VARS']['SC_OPTIONS']['ext/install']['update']['seminars_migrateCategoryIconsToFal']
         = \OliverKlee\Seminars\UpgradeWizards\CategoryIconToFalUpgradeWizard::class;
+    $GLOBALS['TYPO3_CONF_VARS']['SC_OPTIONS']['ext/install']['update']['seminars_migrateSeminarImagesToFal']
+        = \OliverKlee\Seminars\UpgradeWizards\SeminarImageToFalUpgradeWizard::class;
 };
 
 $boot();